events: moved slow event emitter to own file
