+++ lib/events.js
@@ -22,16 +22,11 @@
 'use strict';
 
 const {
-  ArrayPrototypeJoin,
   ArrayPrototypePop,
   ArrayPrototypePush,
-  ArrayPrototypeSlice,
-  ArrayPrototypeSplice,
   ArrayPrototypeUnshift,
   Boolean,
   Error,
-  ErrorCaptureStackTrace,
-  FunctionPrototypeBind,
   FunctionPrototypeCall,
   NumberMAX_SAFE_INTEGER,
   ObjectDefineProperties,
@@ -44,8 +39,6 @@ const {
   PromiseResolve,
   ReflectApply,
   ReflectOwnKeys,
-  String,
-  StringPrototypeSplit,
   Symbol,
   SymbolAsyncIterator,
   SymbolDispose,
@@ -55,12 +48,7 @@ const kRejection = SymbolFor('nodejs.rejection');
 
 const { kEmptyObject } = require('internal/util');
 
-const {
-  inspect,
-  identicalSequenceRange,
-} = require('internal/util/inspect');
 
-let spliceOne;
 let FixedQueue;
 let kFirstEventParam;
 let kResistStopPropagation;
@@ -70,10 +58,7 @@ const {
   codes: {
     ERR_INVALID_ARG_TYPE,
     ERR_INVALID_THIS,
-    ERR_UNHANDLED_ERROR,
   },
-  genericNodeError,
-  kEnhanceStackBeforeInspector,
 } = require('internal/errors');
 
 const {
@@ -86,20 +71,27 @@ const {
   validateString,
 } = require('internal/validators');
 const { addAbortListener } = require('internal/events/abort_listener');
-const { kInitialFastPath } = require('internal/events/symbols');
+const {
+  kInitialFastPath,
+  kCapture,
+  kErrorMonitor,
+  kShapeMode,
+  kMaxEventTargetListeners,
+  kMaxEventTargetListenersWarned,
+  kWatermarkData,
+  kImpl,
+  kIsFastPath,
+  kSwitchToSlowPath,
+  kEvents,
+} = require('internal/events/symbols');
+const {
+  arrayClone,
+  addCatch,
+  throwErrorOnMissingErrorHandler,
+} = require('internal/events/shared_internal_event_emitter');
+const { SlowEventEmitter } = require('internal/events/slow_event_emitter');
+// Const { } = require('internal/events/fast_event_emitter');
 
-
-const kCapture = Symbol('kCapture');
-const kErrorMonitor = Symbol('events.errorMonitor');
-const kShapeMode = Symbol('shapeMode');
-const kMaxEventTargetListeners = Symbol('events.maxEventTargetListeners');
-const kMaxEventTargetListenersWarned =
-  Symbol('events.maxEventTargetListenersWarned');
-const kWatermarkData = SymbolFor('nodejs.watermarkData');
-
-const kImpl = Symbol('kImpl');
-const kIsFastPath = Symbol('kIsFastPath');
-const kSwitchToSlowPath = Symbol('kSwitchToSlowPath');
 
 let EventEmitterAsyncResource;
 // The EventEmitterAsyncResource has to be initialized lazily because event.js
@@ -443,7 +435,7 @@ EventEmitter.setMaxListeners =
 EventEmitter.init = function(opts) {
   // TODO(rluvaton): remove this eslint ignore
   // eslint-disable-next-line no-use-before-define
-  this[kImpl] ??= new FastEventEmitter(this);
+  this[kImpl] ??= new FastEventEmitter(this, opts?.[kEvents]);
   this[kIsFastPath] ??= true;
 
   const thisEvents = this._events;
@@ -520,30 +512,6 @@ EventEmitter.prototype.getMaxListeners = function getMaxListeners() {
   return _getMaxListeners(this);
 };
 
-function enhanceStackTrace(err, own) {
-  let ctorInfo = '';
-  try {
-    const { name } = this.constructor;
-    if (name !== 'EventEmitter')
-      ctorInfo = ` on ${name} instance`;
-  } catch {
-    // Continue regardless of error.
-  }
-  const sep = `\nEmitted 'error' event${ctorInfo} at:\n`;
-
-  const errStack = ArrayPrototypeSlice(
-    StringPrototypeSplit(err.stack, '\n'), 1);
-  const ownStack = ArrayPrototypeSlice(
-    StringPrototypeSplit(own.stack, '\n'), 1);
-
-  const { len, offset } = identicalSequenceRange(ownStack, errStack);
-  if (len > 0) {
-    ArrayPrototypeSplice(ownStack, offset + 1, len - 2,
-                         '    [... lines matching original stack trace ...]');
-  }
-
-  return err.stack + sep + ArrayPrototypeJoin(ownStack, '\n');
-}
 
 /**
  * Synchronously calls each of the listeners registered
@@ -844,19 +812,6 @@ function listenerCount(type, listener) {
 EventEmitter.prototype.eventNames = function eventNames() {
   return this._eventsCount > 0 ? ReflectOwnKeys(this._events) : [];
 };
-
-function arrayClone(arr) {
-  // At least since V8 8.3, this implementation is faster than the previous
-  // which always used a simple for-loop
-  switch (arr.length) {
-    case 2: return [arr[0], arr[1]];
-    case 3: return [arr[0], arr[1], arr[2]];
-    case 4: return [arr[0], arr[1], arr[2], arr[3]];
-    case 5: return [arr[0], arr[1], arr[2], arr[3], arr[4]];
-    case 6: return [arr[0], arr[1], arr[2], arr[3], arr[4], arr[5]];
-  }
-  return ArrayPrototypeSlice(arr);
-}
 
 function unwrapListeners(arr) {
   const ret = arrayClone(arr);
@@ -1323,349 +1278,6 @@ class FastEventEmitter {
 
 // --------------- SLOW
 
-
-// TODO - should have the same API as slow_event_emitter
-
-class SlowEventEmitter {
-  _events = undefined;
-  _eventsCount = 0;
-  // TODO - convert to symbol and rename
-  eventEmitterTranslationLayer = undefined;
-  [kShapeMode] = undefined;
-
-  /**
-   *
-   * @param {FastEventEmitter} fastEventEmitter
-   */
-  static fromFastEventEmitter(fastEventEmitter) {
-    const eventEmitter = new SlowEventEmitter(fastEventEmitter.eventEmitterTranslationLayer, fastEventEmitter._events);
-
-    // TODO - add missing
-    eventEmitter._eventsCount = fastEventEmitter._eventsCount;
-    eventEmitter[kShapeMode] = fastEventEmitter[kShapeMode];
-
-    return eventEmitter;
-  }
-
-  constructor(eventEmitterTranslationLayer, events = undefined) {
-    this.eventEmitterTranslationLayer = eventEmitterTranslationLayer;
-    this._events = events;
-  }
-
-  /**
-   * Synchronously calls each of the listeners registered
-   * for the event.
-   * @param {string | symbol} type
-   * @param {...any} [args]
-   * @returns {boolean}
-   */
-  emit(type, ...args) {
-    let doError = (type === 'error');
-    const eventEmitterTranslationLayer = this.eventEmitterTranslationLayer;
-
-    const events = this._events;
-    if (events !== undefined) {
-      if (doError && events[kErrorMonitor] !== undefined)
-        this.emit(kErrorMonitor, ...args);
-      doError = (doError && events.error === undefined);
-    } else if (!doError)
-      return false;
-
-    // If there is no 'error' event listener then throw.
-    if (doError) {
-      throwErrorOnMissingErrorHandler.apply(eventEmitterTranslationLayer, args);
-    }
-
-    const handler = events[type];
-
-    if (handler === undefined)
-      return false;
-
-    if (typeof handler === 'function') {
-      const result = handler.apply(eventEmitterTranslationLayer, args);
-
-      // We check if result is undefined first because that
-      // is the most common case so we do not pay any perf
-      // penalty
-      if (result !== undefined && result !== null) {
-        addCatch(eventEmitterTranslationLayer, result, type, args);
-      }
-    } else {
-      const len = handler.length;
-      const listeners = arrayClone(handler);
-      for (let i = 0; i < len; ++i) {
-        const result = listeners[i].apply(eventEmitterTranslationLayer, args);
-
-        // We check if result is undefined first because that
-        // is the most common case so we do not pay any perf
-        // penalty.
-        // This code is duplicated because extracting it away
-        // would make it non-inlineable.
-        if (result !== undefined && result !== null) {
-          addCatch(eventEmitterTranslationLayer, result, type, args);
-        }
-      }
-    }
-
-    return true;
-  }
-
-  /**
-   * Adds a listener to the event emitter.
-   * @param {string | symbol} type
-   * @param {Function} listener
-   * @returns {EventEmitter}
-   */
-  addListener(type, listener, prepend) {
-    const target = this.eventEmitterTranslationLayer;
-    let m;
-    let events;
-    let existing;
-
-    events = this._events;
-    if (events === undefined) {
-      events = this._events = { __proto__: null };
-      this._eventsCount = 0;
-    } else {
-      // To avoid recursion in the case that type === "newListener"! Before
-      // adding it to the listeners, first emit "newListener".
-      if (events.newListener !== undefined) {
-        this.eventEmitterTranslationLayer.emit('newListener', type,
-                                               listener.listener ?? listener);
-
-        // Re-assign `events` because a newListener handler could have caused the
-        // this._events to be assigned to a new object
-        events = this._events;
-      }
-      existing = events[type];
-    }
-
-    if (existing === undefined) {
-      // Optimize the case of one listener. Don't need the extra array object.
-      events[type] = listener;
-      ++this._eventsCount;
-    } else {
-      if (typeof existing === 'function') {
-        // Adding the second element, need to change to array.
-        existing = events[type] =
-          prepend ? [listener, existing] : [existing, listener];
-        // If we've already got an array, just append.
-      } else if (prepend) {
-        existing.unshift(listener);
-      } else {
-        existing.push(listener);
-      }
-
-      // Check for listener leak
-      m = _getMaxListeners(target);
-      if (m > 0 && existing.length > m && !existing.warned) {
-        existing.warned = true;
-        // No error code for this since it is a Warning
-        const w = genericNodeError(
-          `Possible EventEmitter memory leak detected. ${existing.length} ${String(type)} listeners ` +
-          `added to ${inspect(target, { depth: -1 })}. Use emitter.setMaxListeners() to increase limit`,
-          { name: 'MaxListenersExceededWarning', emitter: target, type: type, count: existing.length });
-        process.emitWarning(w);
-      }
-    }
-
-    return this;
-  }
-
-  /**
-   * Removes the specified `listener`.
-   * @param {string | symbol} type
-   * @param {Function} listener
-   * @returns {EventEmitter}
-   */
-  removeListener(type, listener) {
-    const events = this._events;
-    if (events === undefined)
-      return undefined;
-
-    const list = events[type];
-    if (list === undefined)
-      return undefined;
-
-    if (list === listener || list.listener === listener) {
-      this._eventsCount -= 1;
-
-      if (this[kShapeMode]) {
-        events[type] = undefined;
-      } else if (this._eventsCount === 0) {
-        this._events = { __proto__: null };
-      } else {
-        delete events[type];
-        if (events.removeListener)
-          this.emit('removeListener', type, list.listener || listener);
-      }
-    } else if (typeof list !== 'function') {
-      let position = -1;
-
-      for (let i = list.length - 1; i >= 0; i--) {
-        if (list[i] === listener || list[i].listener === listener) {
-          position = i;
-          break;
-        }
-      }
-
-      if (position < 0)
-        return this;
-
-      if (position === 0)
-        list.shift();
-      else {
-        if (spliceOne === undefined)
-          spliceOne = require('internal/util').spliceOne;
-        spliceOne(list, position);
-      }
-
-      if (list.length === 1)
-        events[type] = list[0];
-
-      if (events.removeListener !== undefined)
-        this.emit('removeListener', type, listener);
-    }
-
-    return undefined;
-  }
-
-  /**
-   * Removes all listeners from the event emitter. (Only
-   * removes listeners for a specific event name if specified
-   * as `type`).
-   * @param {string | symbol} [type]
-   * @returns {EventEmitter}
-   */
-  removeAllListeners(type) {
-    const events = this._events;
-    if (events === undefined)
-      return undefined;
-
-    // Not listening for removeListener, no need to emit
-    if (events.removeListener === undefined) {
-      if (arguments.length === 0) {
-        this._events = { __proto__: null };
-        this._eventsCount = 0;
-      } else if (events[type] !== undefined) {
-        if (--this._eventsCount === 0)
-          this._events = { __proto__: null };
-        else
-          delete events[type];
-      }
-      this[kShapeMode] = false;
-      return undefined;
-    }
-
-    // Emit removeListener for all listeners on all events
-    if (arguments.length === 0) {
-      for (const key of ReflectOwnKeys(events)) {
-        if (key === 'removeListener') continue;
-        this.removeAllListeners(key);
-      }
-      this.removeAllListeners('removeListener');
-      this._events = { __proto__: null };
-      this._eventsCount = 0;
-      this[kShapeMode] = false;
-      return undefined;
-    }
-
-    const listeners = events[type];
-
-    if (typeof listeners === 'function') {
-      this.removeListener(type, listeners);
-    } else if (listeners !== undefined) {
-      // LIFO order
-      for (let i = listeners.length - 1; i >= 0; i--) {
-        this.removeListener(type, listeners[i]);
-      }
-    }
-
-    return undefined;
-  }
-}
-
-
-function addCatch(that, promise, type, args) {
-  if (!that[kCapture]) {
-    return;
-  }
-
-  // Handle Promises/A+ spec, then could be a getter
-  // that throws on second use.
-  try {
-    const then = promise.then;
-
-    if (typeof then === 'function') {
-      then.call(promise, undefined, function(err) {
-        // The callback is called with nextTick to avoid a follow-up
-        // rejection from this promise.
-        process.nextTick(emitUnhandledRejectionOrErr, that, err, type, args);
-      });
-    }
-  } catch (err) {
-    that.emit('error', err);
-  }
-}
-
-function emitUnhandledRejectionOrErr(ee, err, type, args) {
-  if (typeof ee.eventEmitterTranslationLayer[kRejection] === 'function') {
-    ee.eventEmitterTranslationLayer[kRejection](err, type, ...args);
-  } else {
-    // We have to disable the capture rejections mechanism, otherwise
-    // we might end up in an infinite loop.
-    const prev = ee.eventEmitterTranslationLayer[kCapture];
-
-    // If the error handler throws, it is not catchable and it
-    // will end up in 'uncaughtException'. We restore the previous
-    // value of kCapture in case the uncaughtException is present
-    // and the exception is handled.
-    try {
-      ee.eventEmitterTranslationLayer[kCapture] = false;
-      ee.emit('error', err);
-    } finally {
-      ee.eventEmitterTranslationLayer[kCapture] = prev;
-    }
-  }
-}
+// ------ todo
 
 // ------- Internal helpers
-
-
-// TODO - move this to a different file
-// TODO - rename
-function throwErrorOnMissingErrorHandler(...args) {
-  let er;
-  if (args.length > 0)
-    er = args[0];
-
-  if (er instanceof Error) {
-    try {
-      const capture = {};
-      ErrorCaptureStackTrace(capture, EventEmitter.prototype.emit);
-      ObjectDefineProperty(er, kEnhanceStackBeforeInspector, {
-        __proto__: null,
-        value: FunctionPrototypeBind(enhanceStackTrace, this, er, capture),
-        configurable: true,
-      });
-    } catch {
-      // Continue regardless of error.
-    }
-
-    // Note: The comments on the `throw` lines are intentional, they show
-    // up in Node's output if this results in an unhandled exception.
-    throw er; // Unhandled 'error' event
-  }
-
-  let stringifiedEr;
-  try {
-    stringifiedEr = inspect(er);
-  } catch {
-    stringifiedEr = er;
-  }
-
-  // At least give some kind of context to the user
-  const err = new ERR_UNHANDLED_ERROR(stringifiedEr);
-  err.context = er;
-  throw err; // Unhandled 'error' event
-}
