+++ lib/internal/events/fast_event_emitter.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = {};
