+++ lib/internal/events/symbols.js
@@ -2,12 +2,34 @@
 
 const {
   Symbol,
+  SymbolFor,
 } = primordials;
 
 const kFirstEventParam = Symbol('nodejs.kFirstEventParam');
 const kInitialFastPath = Symbol('kInitialFastPath');
 
+const kCapture = Symbol('kCapture');
+const kErrorMonitor = Symbol('events.errorMonitor');
+const kShapeMode = Symbol('shapeMode');
+const kMaxEventTargetListeners = Symbol('events.maxEventTargetListeners');
+const kMaxEventTargetListenersWarned =
+  Symbol('events.maxEventTargetListenersWarned');
+const kWatermarkData = SymbolFor('nodejs.watermarkData');
+
+const kImpl = Symbol('kImpl');
+const kIsFastPath = Symbol('kIsFastPath');
+const kSwitchToSlowPath = Symbol('kSwitchToSlowPath');
+
 module.exports = {
   kFirstEventParam,
   kInitialFastPath,
+  kCapture,
+  kErrorMonitor,
+  kShapeMode,
+  kMaxEventTargetListeners,
+  kMaxEventTargetListenersWarned,
+  kWatermarkData,
+  kImpl,
+  kIsFastPath,
+  kSwitchToSlowPath,
 };
